Configure SRPs and update bounds for 11.2 integration

+++ cardano-testnet/src/Testnet/Blockfrost.hs
@@ -273,7 +273,7 @@ blockfrostToGenesis (shelleyGenesis', alonzoGenesis', conwayGenesis', dijkstraGe
       , sppExtraEntropy = bfgExtraEntropy
       , sppProtocolVersion = ProtVer
         { pvMajor = bfgProtocolMajorVer
-        , pvMinor = bfgProtocolMinorVer
+        , pvMinor = fromIntegral bfgProtocolMinorVer
         }
       , sppMinUTxOValue = compactCoinOrError bfgMinUTxO
       , sppMinPoolCost = compactCoinOrError bfgMinPoolCost
